Validate rules docs when running unit tests
Fix #1734
